We need to use "or" in case of multiple metrics
Otherwise, we have to have machine and other fields that have all metrics. This is not the case when, for example, tests were executed in different configurations.

The change affects only the cases when metrics are filters for an initial query and this is a case in metrics mode only.